Rename store file
