+++ src/actions/makeActionCreator.js
@@ -1,5 +1,5 @@
 export const makeActionCreator = actionType => {
     return payload => {
-        return { type: actionType, payload: payload };
+        return { type: actionType, payload };
     };
 };
