+++ src/ducks/organization/reducer.js
@@ -1,12 +1,12 @@
 import update from 'immutability-helper';
-import initialState from 'store/store';
+import initialState from 'store/initialState';
 
 import {
     ADD_ORGANIZATION,
     SET_ORGANIZATION
 } from './types';
 
-const organizationReducer = (state = initialState, action) => {
+const organizationReducer = (state = initialState.organization, action) => {
     const { type, payload } = action;
     switch (type) {
         case SET_ORGANIZATION:
@@ -15,7 +15,7 @@ const organizationReducer = (state = initialState, action) => {
                     $set: payload
                 }
             });
-        // TODO finish this case
+        // TODO: finish this case
         case ADD_ORGANIZATION:
             return state; // For now
         default:
