+++ src/store/initialState.js
@@ -12,6 +12,14 @@ const initialState = {
         selectedOrg: '',
     },
     cart: JSON.parse(localStorage.getItem('cart')) || [],
+    pageData: {
+        pageName: null,
+        description: 'Fundraising platform for non-profits and charities',
+        text: 'Fundraising platform for non-profits and charities',
+        image: require('../Assets/Logo/PNG/Artboard-1-copy-2Generosity-Logo.png'),
+        title: 'Generosity Market',
+        url: window.location.href,
+    },
     // filter: 'all',
     // alert: { type: null, message: null }
 };
